Move defir_to_gtir to frontend
