+++ .pre-commit-config.yaml
@@ -114,7 +114,7 @@ repos:
       src/gt4py/frontend/__init__.py |
       src/gt4py/frontend/gtscript_frontend.py |
       src/gt4py/frontend/nodes.py |
-      src/gt4py/frontend/utils.py |
+      src/gt4py/frontend/node_util.py |
       src/gt4py/storage/__init__.py |
       src/gt4py/storage/storage.py |
       src/gt4py/storage/utils.py |
@@ -153,7 +153,7 @@ repos:
       src/gt4py/backend/pyext_builder.py |
       src/gt4py/frontend/gtscript_frontend.py |
       src/gt4py/frontend/nodes.py |
-      src/gt4py/frontend/utils.py |
+      src/gt4py/frontend/node_util.py |
       src/gt4py/utils/meta.py |
       tests/definitions.py |
       tests/eve_tests/definitions.py |
