+++ src/gt4py/frontend/node_util.py
